Done after God gave me some wisdom and brought me to sky full of white lights

+++ src/main/java/ua/edu/ucu/stream/AsIntStream.java
@@ -130,7 +130,7 @@ public class AsIntStream implements IntStream {
                 if (value == null) {
                     throw new NoSuchElementException();
                 }
-                return value.intValue();
+                return value;
             }
         }
